servers/mcp-verifica-codici/pyproject.toml
[project]
name = "verifica_codici"
version = "0.1.0"
description = "Un server MCP per la validazione incrociata dei codici di documenti di progetto."
requires-python = "~=3.12"

dependencies = [
    # Dipendenze comuni per MCP e Pydantic
    "mcp~=1.1.3",
    "anyio",
    "sniffio",
    "idna",
    "jsonschema",
    "jsonschema-specifications",
    "referencing",
    "rpds-py",
    "attrs",
    "pydantic==2.8.2",
    "pydantic_core==2.20.1",
    "annotated-types",
    "typing_extensions",

    # === DIPENDENZE SPECIFICHE PER QUESTO TOOL ===
    # Per l'analisi dei PDF
    "pdfplumber",

    # Per l'elaborazione delle immagini
    "opencv-python",
    "numpy",
    "Pillow",

    # Per l'OCR
    "pytesseract",

    # Per le chiamate all'AI e la gestione delle chiavi
    "openai",
    "python-dotenv",
    "aiohttp"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"